requirements.txt
Flask==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
json5==0.9.14
Markdown==3.5
requests==2.31.0
requests-oauthlib==1.3.1